requirements.txt
python-dotenv==1.0.0
requests==2.31.0
Flask==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
